Initial Dataset
This dataset was taken from https://www.fhfa.gov/PolicyProgramsResearch/Research/PaperDocuments/bdl_faqs_local_hpis.pdf and is considered experimental.  

The initial dataset has transactions involving conforming, conventional mortgages
that are purchased or securitized by Fannie Mae or Freddie Mac. Only singlefamily homes are included (condominiums, cooperatives, multi-unit properties,
and planned unit developments are excluded). We use a statistical method to
measure the annual price change in repeat sales or refinancings on the same
property with data that extend back to the mid-1970s. The underlying information
is the same as used in the standard “all-transactions” FHFA HPIs and is based on
almost 100 million observations.

Further details on how the data was collected and assessed are available via the above URL.